Updated Dot movement to be smoother
Updated Dot movement to use Key States instead of Key Events.  This
necessitated the removal of the handle_input method, and I changed the
move method to be called update, as it is not more universal.

This is also getting ready for the first RakNet test, where the Dots are
networked.